stuff2

+++ layer/VkLayer_FROG_gamescope_wsi.cpp
@@ -985,20 +985,20 @@ namespace GamescopeWSILayer {
             continue;
           }
           
-          static std::valarray<uint64_t> durations(32);
+          static std::valarray<int64_t> durations(32);
           static uint64_t counter = 0; 
 		  
 		  uint64_t start = get_time_in_nanos();
           const bool canBypass = gamescopeSurface->canBypassXWayland();
-          durations[counter]=(get_time_in_nanos()-start);
+          durations[counter]=static_cast<int64_t>(get_time_in_nanos()-start);
           static constexpr float nsPerMs = 1'000'000.0;
           fprintf(stderr, "canBypassXWayland(): %.2fms\n",  (durations[counter])/nsPerMs);
           if (++counter == 32) {
           	counter=0;
-          	uint64_t mean = ( (durations.sum()) / (32ul) );
+          	int64_t mean = ( (durations.sum()) / (32l) );
           	std::adjacent_difference(std::begin(durations), std::end(durations), std::begin(durations));
-          	const std::valarray<uint64_t> selectTheseIndices = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31};
-          	const std::valarray<uint64_t> selection = durations[selectTheseIndices];
+          	const std::valarray<int64_t> selectTheseIndices = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31};
+          	const std::valarray<int64_t> selection = durations[selectTheseIndices];
           	auto [min, max] = std::minmax_element(std::begin(selection), std::end(selection));
           	
           	fprintf(stderr, "\n\n\n\ncanBypassXWayland() average duration:%.2fms,\
